fix(rpc/offline): adjust_stock integer cast + surface drain errors
## Two bugs found in production after the Phase 1 RPC migrations landed.

### Bug 1 — adjust_stock signature mismatch

`adjust_stock(p_id bigint, qty_delta integer, p_reason text, p_ref_table
text, p_ref_id bigint)` expects qty_delta as **integer** but both
`create_sale_order_with_items` and `create_stock_movement_with_items`
were passing `(item->>'quantity')::numeric`. With Postgres named-arg
dispatch this throws:

  function public.adjust_stock(p_id => bigint, qty_delta => numeric,
  p_reason => unknown, p_ref_table => unknown, p_ref_id => bigint)
  does not exist

Every sale, receive, claim, and return failed.

Fix: cast to ::integer at the call site. Items.quantity columns are all
integer in the schema anyway, so this is a strict identity, no semantic
change. Already applied to the live DB via apply_migration; the local
SQL files in supabase-migrations/ are now in sync for any future fresh
deploys.

### Bug 2 — offline queue stuck on a hard error

When Bug 1 fired, the offline queue's drainQueue() caught the error and
broke out of the loop (correct — don't infinite-retry a malformed
payload), but the OfflineBanner had no way to know. It kept showing
"กำลัง sync…" forever even though no sync was in flight.

Fix:
- offline-queue.js exposes a new drain-state pub/sub
  (onDrainStateChange / getDrainState) with three states:
    idle / draining / error  + lastError + lastDrainAt
- drainQueue() now accepts an optional `mapErr` (we pass mapError() so
  the message renders in Thai) and updates the state.
- New `deleteQueuedSale(id)` escape hatch for manually clearing a
  permanently malformed payload.
- OfflineBanner now has THREE visual states:
    red       offline (will queue)
    yellow    online + draining or queue has pending items
    dark red  online + queue stuck on a hard error
              → shows the Thai error and a "ลองส่งอีกครั้ง" button
- main.jsx exposes the new helpers on `window` so the cashier (or a
  console-savvy admin) can debug from DevTools.

After this lands the user's stuck "กำลัง sync…" banner will:
  1. drain on next focus / online event (Bug 1 is fixed in DB)
  2. if it still fails for any reason, surface the actual Thai error +
     a retry button — never silent again.

## Verified

- npm test → 33 pass.
- npm run build → clean. dist/index.html + dist/sw.js produced.

## What user must do after deploy

Reload the POS once (Ctrl+Shift+R). The queued offline sale that was
stuck should now drain automatically. If the banner shows the dark-red
"ส่งบิลในคิวไม่สำเร็จ…" UI instead, click "ลองส่งอีกครั้ง" — the queue
should clear.

+++ supabase-migrations/001_create_sale_order_with_items.sql
@@ -71,7 +71,7 @@ BEGIN
     v_order_id,
     NULLIF(item->>'product_id','')::bigint,
     item->>'product_name',
-    COALESCE((item->>'quantity')::numeric, 0),
+    COALESCE((item->>'quantity')::integer, 0),
     COALESCE((item->>'unit_price')::numeric, 0),
     COALESCE((item->>'discount1_value')::numeric, 0),
     item->>'discount1_type',
@@ -85,7 +85,10 @@ BEGIN
     IF NULLIF(v_item->>'product_id','') IS NOT NULL THEN
       PERFORM public.adjust_stock(
         p_id        => (v_item->>'product_id')::bigint,
-        qty_delta   => -((v_item->>'quantity')::numeric),
+        -- qty_delta MUST be cast to integer; adjust_stock(p_id bigint,
+        -- qty_delta integer, ...) and Postgres named-arg dispatch does
+        -- not implicit-cast numeric → integer.
+        qty_delta   => -((v_item->>'quantity')::integer),
         p_reason    => 'sale',
         p_ref_table => 'sale_orders',
         p_ref_id    => v_order_id
